requirements.txt
demjson3
duckduckgo_search
googlesearch-python
langchain-core~=0.3.51
langchain-openai==0.3.12
langgraph==0.3.24
lxml
lxml_html_clean
markdownify
numpy
pandas==2.2.3
playwright
pydantic~=2.11.2
PyPDF2
pytest
python-dotenv==1.1.0
readability-lxml
requests
scikit-learn==1.6.1
tavily-python
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124

